Rename userspace proxy to udp proxy
